feat: Move messages to the public folder to allow editing post-build
