+++ src/UI/public/messages/fr-template.json
