+++ src/UI/public/messages/fr.json
