+++ src/UI/utils/localization.ts
@@ -26,6 +26,20 @@ export type TranslationMessage = {
   pt: string;
 };
 
+const isProduction = process.env.NODE_ENV === 'production';
+const messagesDirectory = isProduction
+  ? '../../standalone/messages'
+  : '../messages';
+console.log('Is PROD: ', isProduction);
+
+/**
+ * Ensure the messages JSON file is placed in the public directory, it will be served as a static asset.
+ * This will enable users to modify the labels directly.
+ * If the JSON file is not in the public directory, it's typically intended for server-side use or to be bundled with your application. This will
+ * ensure the JSON file cannot be modified
+ * @param context
+ * @returns
+ */
 export const getMessages = async (context: GetServerSidePropsContext) => {
   const locale = context.locale || 'en';
   const cookies = context.req.cookies;
@@ -33,7 +47,10 @@ export const getMessages = async (context: GetServerSidePropsContext) => {
   const cookieLocale = (await cookies['VECTORATLAS_LOCALE']) || 'en';
   return {
     props: {
-      messages: (await import(`../messages/${cookieLocale}.json`)).default,
+      //messages: messages,
+      // messages: (await import(`../messages/${cookieLocale}.json`)).default,
+      messages: (await import(`../public/messages/${cookieLocale}.json`))
+        .default,
       // Note that when `now` is passed to the app, you need to make sure the
       // value is updated from time to time, so relative times are updated. See
       // https://next-intl-docs.vercel.app/docs/usage/configuration#global-now-value
@@ -47,11 +64,8 @@ export const getTranslation = async (key: string, values: any = null) => {
   // this should be up to you, maybe from app state or from cookies
   // const locale = locale() ?? 'en';
   const locale = (store.getState() as AppState).localization.locale || 'en';
-  const messages = (await import(`../messages/${locale}.json`)).default;
-  // messages =
-  //   Object.keys(messages).length == 0
-  //       (await import(`../messages/${locale}.json`)).default
-  //     : messages;
+  const messages = (await import(`${messagesDirectory}/${locale}.json`))
+    .default;
   const t = createTranslator({ locale, messages });
   return t(key, values);
 };
@@ -72,7 +86,8 @@ export const getRawTranslation = async (
   if (language) {
     locale = language;
   }
-  const messages = (await import(`../messages/${locale}.json`)).default;
+  const messages = (await import(`${messagesDirectory}/${locale}.json`))
+    .default;
   console.log('Messages: ', messages);
   // messages =
   //   Object.keys(messages).length == 0
